+++ cgmes-model-interpretation/src/main/java/com/powsybl/cgmes/model/interpretation/AbstractAdmittanceMatrix.java
@@ -1,0 +1,12 @@
+package com.powsybl.cgmes.model.interpretation;
+
+import org.apache.commons.math3.complex.Complex;
+
+abstract class AbstractAdmittanceMatrix {
+
+    Complex     yff;
+    Complex     yft;
+    Complex     ytf;
+    Complex     ytt;
+    DetectedBranchModel branchModel;
+}
