+++ cgmes-model-interpretation/src/main/java/com/powsybl/cgmes/model/interpretation/AbstractAdmittanceMatrix3.java
@@ -1,0 +1,8 @@
+package com.powsybl.cgmes.model.interpretation;
+
+abstract class AbstractAdmittanceMatrix3 {
+
+    AbstractAdmittanceMatrix end1;
+    AbstractAdmittanceMatrix end2;
+    AbstractAdmittanceMatrix end3;
+}
