--- cgmes-validation/src/test/java/com/powsybl/cgmes/validation/test/flow/InterpretationResult.java
@@ -1,8 +1,0 @@
-package com.powsybl.cgmes.validation.test.flow;
-
-public class InterpretationResult {
-    double error;
-    // TODO replace the final report
-    // with the data about the results of the interpretation
-    String report;
-}
