+++ pom.xml
@@ -49,6 +49,7 @@
 		<module>cgmes-validation</module>
 		<module>distribution-core-cgmes</module>
 		<module>cgmes-tools</module>
+		<module>cgmes-model-interpretation</module>
 	</modules>
 	
 	<properties>
